Add minigrep project
This commit adds a replication of the minigrep project. It behaves a bit
like `fgrep`, although it takes no commandline flags.

Observations:

- The lack of `if` is a little confusing at first but the `case` pattern
  is familiar enough to overcome that confusion.
- `main` was rewritten a few times to get the final version. I mostly
  struggled to figure out an elegant way of combining `Result(a,b)`
  values. The `try` keyword is great for this though, preferable over
  the `?` operator from Rust perhaps.

A few differences from the original: errors are not printed to standard
error and the process is not exited with exit code 1. The latter is fine
but not great; generally BEAM languages are used for systems that don't
stop anyway.